sr: rename use-get-absolute-url-if-switch-website to get-absolute-url-if-switch-website
